requirements.txt
Django==1.7
django-enumerify==0.4
djangorestframework==3.3
requests=2.8